AS-30: re-linked register buttons to newRegister, and removed the old register page.
