+++ src/App.js
@@ -14,7 +14,6 @@ import Multiplication from './categoryTypesMath/multiplication.js';
 import Division from './categoryTypesMath/division.js';
 import Fractions from './categoryTypesMath/fractions.js';
 import Level_1 from './question1.js';
-import NewRegister from './newRegister.js';
 import Edit from './components/edit';
 import Create from './components/create';
 import Footer from './footer';
@@ -56,7 +55,6 @@ function App() {
 					<Route exact path='/division' element={<Division />}> </Route>
 					<Route exact path='/fractions' element={<Fractions />}> </Route>
                     <Route exact path='/question1' element={<Level_1 />}> </Route>
-					<Route exact path='/newregister' element={<NewRegister />}> </Route>
 					<Route exact path='/create' element={<Create />}> </Route>
 					<Route exact path='/edit:id' element={<Edit />}> </Route>
 					<Route exact path='/recordList' element={<Record_list />}> </Route>
