+++ src/register.js
@@ -1,59 +1,103 @@
-// import { useState } from "react"
-import './register.css';
-const Register = () => {
-    // const [userName] = useState('')
-    // const [password] = useState('')
+import React, { useState } from "react";
+import { useNavigate } from "react-router";
 
-    // const userRegistration = async (e) =>
-    //     e.preventDefault()
-    // console.log(userName, password)
+export default function Register() {
+  const [form, setForm] = useState({
+    email: "",
+    school: "",
+    password: "",
+  });
+  const navigate = useNavigate();
 
-    return (
-        <>
-        
-  <div className="container">
-    <div className="row">
-      <div className="col-sm-9 col-md-7 col-lg-5 mx-auto">
-        <div className="card border-0 shadow rounded-3 my-5">
-          <div className="card-body p-4 p-sm-5">
-            <h5 className="card-title text-center mb-5 fw-light fs-5">Register</h5>
-            <form>
-              <div className="form-floating mb-3">
-                <input type="email" className="form-control" id="floatingInput" placeholder="name@example.com"></input>
-                <label for="floatingInput">Username</label>
-              </div>
-              <div className="form-floating mb-3">
-                <input type="email" className="form-control" id="floatingInput" placeholder="name@example.com"></input>
-                <label for="floatingInput">School</label>
-              </div>
-              <div className="form-floating mb-3">
-                <input type="password" className="form-control" id="floatingPassword" placeholder="Password"></input>
-                <label for="floatingPassword">Password</label>
-              </div>
-              <div className="form-floating mb-3">
-                <input type="password" className="form-control" id="floatingPassword" placeholder="Password"></input>
-                <label for="floatingPassword">Confirm Password</label>
-              </div>
+  // These methods will update the state properties.
+  function updateForm(value) {
+    return setForm((prev) => {
+      return { ...prev, ...value };
+    });
+  }
 
-              <div className="form-check mb-3">
-                <input className="form-check-input" type="checkbox" value="" id="rememberPasswordCheck"></input>
-                <label className="form-check-label" for="rememberPasswordCheck">
-                  Remember password
-                </label>
+  // This function will handle the submission.
+  async function onSubmit(e) {
+    e.preventDefault();
+
+    // When a post request is sent to the create url, we'll add a new record to the database.
+    const newPerson = { ...form };
+
+    await fetch("http://localhost:5000/login/add", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(newPerson),
+    }).catch((error) => {
+      window.alert(error);
+      return;
+    });
+
+    setForm({ email: "", school: "", password: "" });
+    navigate("/");
+  }
+
+  // This following section will display the form that takes the input from the user.
+  return (
+    <>
+      <div className="container">
+        <div className="row">
+          <div className="col-sm-9 col-md-7 col-lg-5 mx-auto">
+            <div className="card border-0 shadow rounded-3 my-5">
+              <div className="card-body p-4 p-sm-5">
+                <h5 className="card-title text-center mb-5 fw-light fs-5">
+                  Register
+                </h5>
+                <form onSubmit={onSubmit}>
+                  <div className="form-floating mb-3">
+                    <input
+                      type="text"
+                      className="form-control"
+                      id="email"
+                      placeholder="name@example.com"
+                      value={form.email}
+                      onChange={(e) => updateForm({ email: e.target.value })}
+                    ></input>
+                    <label htmlFor="email">Username/Email</label>
+                  </div>
+                  <div className="form-floating mb-3">
+                    <input
+                      type="text"
+                      className="form-control"
+                      id="school"
+                      placeholder="Oakland Elementary School"
+                      value={form.school}
+                      onChange={(e) => updateForm({ school: e.target.value })}
+                    ></input>
+                    <label htmlFor="school">School</label>
+                  </div>
+                  <div className="form-floating mb-3">
+                    <input
+                      type="password"
+                      className="form-control"
+                      id="password"
+                      placeholder="Password"
+                      value={form.password}
+                      onChange={(e) => updateForm({ password: e.target.value })}
+                    ></input>
+                    <label htmlFor="password">Password</label>
+                  </div>
+                  <div className="d-grid">
+                    <button
+                      className="btn btn-primary btn-login text-uppercase fw-bold"
+                      value="Submit"
+                      type="submit"
+                    >
+                      Sign in
+                    </button>
+                  </div>
+                </form>
               </div>
-              <div className="d-grid">
-                <button className="btn btn-primary btn-login text-uppercase fw-bold" type="submit">Sign
-                  in</button>
-              </div>
-              
-            </form>
+            </div>
           </div>
         </div>
       </div>
-    </div>
-  </div>
-
-</>
-    )
+    </>
+  );
 }
-export default Register
